+++ fastapi_limiter/depends.py
@@ -39,4 +39,4 @@ class RateLimiter:
         if num == 1:
             await redis.pexpire(key, self.milliseconds)
         if num > self.times:
-            return await callback(request, pexpire)
+            return await callback(request, response, pexpire)