+++ examples/main.py
@@ -19,5 +19,16 @@ async def index():
     return {"msg": "Hello World"}
 
 
+@app.get(
+    "/multiple",
+    dependencies=[
+        Depends(RateLimiter(times=1, seconds=5)),
+        Depends(RateLimiter(times=2, seconds=15)),
+    ],
+)
+async def multiple():
+    return {"msg": "Hello World"}
+
+
 if __name__ == "__main__":
     uvicorn.run("main:app", debug=True, reload=True)